LPS-107468 Use consistent, longer timeouts in Karma test runs
As per this comment:

https://github.com/brianchandotcom/liferay-portal/pull/83566#issuecomment-576797790

This is another example of an unstable test, like the one we ameliorated
in c652366bf373ed12f224eb50a5bcc7de6c1047a0. The related console output
is:

> timeout of 2000ms exceeded. Ensure the done() callback is being called
> in this test.

In this commit I am raising the timeouts across the board in the 3
places where we are currently using Karma. I'm not putting this config
in the "liferay-karma-config" package:

https://www.npmjs.com/package/liferay-karma-config

because it is in a non-org repo, was last published 3 years ago, and we
don't think we actually want to encourage further use of Karma precisely
because of these kinds of stability issues.

Note that in addition to lifting the Mocha timeout, which is noted
as defaulting to 2000ms here:

http://blog.jonathanargentiero.com/overriding-karma-mocha-default-timeout-2000ms/

I'm also lifting the `browserDisconnectTimeout`, because that's the
other potential culprit here, as the only other possible configuration
value that defaults to 2000ms:

http://karma-runner.github.io/4.0/config/configuration-file.html